Reuse the same code for profile connections
