+++ frontend/src/lib/schema.d.ts
@@ -2894,16 +2894,14 @@ export interface operations {
     };
     otodb_api_profile_edit_connections: {
         parameters: {
-            query?: never;
+            query: {
+                urls: string;
+            };
             header?: never;
             path?: never;
             cookie?: never;
         };
-        requestBody: {
-            content: {
-                "application/json": components["schemas"]["ConnectionSchema"][];
-            };
-        };
+        requestBody?: never;
         responses: {
             /** @description OK */
             200: {
